Firebase module refactor
