+++ src/modules/authentication/components/index.ts
@@ -2,4 +2,3 @@ export * from './AuthenticationLayout';
 export * from './AuthenticationButtons';
 export * from './AuthenticationForm';
 export * from './SideRandomImage';
-export * from './FirebaseAuthProvider';
