--- src/modules/firebase/store/firebase.ts
@@ -1,34 +1,0 @@
-import { initializeApp } from 'firebase/app';
-import {
-  FacebookAuthProvider,
-  getAuth,
-  GoogleAuthProvider,
-} from 'firebase/auth';
-import { enableIndexedDbPersistence, getFirestore } from 'firebase/firestore';
-import { getFunctions } from 'firebase/functions';
-import { getStorage } from 'firebase/storage';
-
-export const firebaseApp = initializeApp({
-  apiKey: process.env.REACT_APP_FIREBASE_API_KEY,
-  authDomain: process.env.REACT_APP_FIREBASE_AUTH_DOMAIN,
-  projectId: process.env.REACT_APP_FIREBASE_PROJECT_ID,
-  storageBucket: process.env.REACT_APP_FIREBASE_STORAGE_BUCKET,
-  messagingSenderId: process.env.REACT_APP_FIREBASE_SENDER_ID,
-  appId: process.env.REACT_APP_FIREBASE_APP_ID,
-});
-
-export const functions = getFunctions(firebaseApp);
-export const db = getFirestore(firebaseApp);
-export const storage = getStorage(firebaseApp);
-export const auth = getAuth(firebaseApp);
-export const googleProvider = new GoogleAuthProvider();
-export const facebookProvider = new FacebookAuthProvider();
-googleProvider.setCustomParameters({ prompt: 'select_account' });
-
-enableIndexedDbPersistence(db, { forceOwnership: true }).catch(err => {
-  if (err.code == 'failed-precondition') {
-    console.log('error code', err);
-  } else if (err.code == 'unimplemented') {
-    console.log('error code', err);
-  }
-});
